Lamps count calculations

+++ LaborProtection.Services/LightServices/ILightService.cs
@@ -7,7 +7,7 @@ namespace LaborProtection.Services.LightServices
 {
     public interface ILightService
     {
-        double GetLampCount(RoomEntity roomEntity, LampEntity lampEntity, BulbEntity bulbEntity, double floorReflection, double wallReflection, double ceillingReflection, LampType lampType);
+        int GetLampCount(RoomEntity roomEntity, LampEntity lampEntity, BulbEntity bulbEntity, double floorReflection, double wallReflection, double ceillingReflection, LampType lampType);
         double LightFlux(RoomEntity roomEntity, LampEntity lampEntity, double floorReflection, double wallReflection, double ceillingReflection, LampType lampType);
         double RoomIndex(RoomEntity roomEntity, LampEntity lampEntity);
         double LampHeightSuspension(double roomHeight, double lampHeight, double tableHeight);
